Fix long-standing GethashCode x64 early halt bug
As documented in connect bug 519104, String.GetHashCode ignores any characters in the string beyond the first null character in x64 runtime. This means that if someone embeds a NULL character in a string, hash-code will only be computed up to that character. This is wrong in that Strings are Length specified and embedded NULL character **are** significant and characters after it **are** important to the hash-code.

NOTE:
1) The return value in 32-bit builds DOES take all characters into account, so the strings that have a nice distribution in x32 builds might generate horrific collisions in x64 builds.
2) While as an implementation detail the String buffer is always followed by a NULL character, this is merely to ease the interop issues and has NOTHING to do with how strings should be handled in pure CLR.
3) Changing the hash-code in a new build is NOT a breaking change, just look at the comment in line 892-896.