Fix unused-private-field warning
The warning is fixed mostly by adding __attribute__((unused)) to the
unused private fields. In most cases, these fields are either paddings or
members of structs / classes that have a managed counterpart and
need to match its layout or seems to serve for debugging purposes.
Several places are fixed by conditionally ifdefing-out the members.

For DAC build though, the warning is left disabled since a large
amount of classes have their members unused in the DAC build, which
is ok.