Fix SIGINT and SIGQUIT handling on Linux / FreeBSD
This change changes the way the SIGINT and SIGQUIT are handled. First, it
makes sure that when the signal is not handled, we call the default signal
handler that takes care of the proper default behavior.
And second, it makes debugging experience better. Before, we had a thread
that was waiting for these signals and then spawning a console event handling
thread. That caused the signals to be detected in that thread even when
running under the debugger, effectively making ctrl-c break-in unusable, since
an attempt to continue after the ctrl-c caused the process to exit.
I've changed it so that the SIGINT and SIGQUIT call signal handlers and those
handlers then spawn a helper thread to handle the console event.
This enables the debugger to take over the signal handling properly.
I had to use a plain pthread insted of creating a PAL thread to make sure
there is no danger of a deadlock in case the original thread was interrupted
while holding a lock that the PAL thread creation use.