Change order in which P/Invoke names are checked
In #1248 we added support for automatically prepending a prefix
onto a DllImport name, in addition to the existing support for
automatically appending a suffix.  This leads to four combinations
that are checked: name, prefix+name, name+suffix, and
prefix+name+suffix.  The current code as of #1248 first checks
for prefix+name+suffix, under the assumption that it'll be the
most common case, and thus we optimize for it, e.g. a developer
writes "sqlite3" such that it'll become "sqlite3.dll" on Windows,
"libsqlite3.so" on Linux, "libsqlite3.dylib" on OS X, etc.

However, there is a debate about whether we should actually check
first for exactly what the developer wrote, checking for just
name before checking for prefix+name+suffix.  Whereas the argument
for prefix+name+suffix is performance, the argument for name is
safety, in that by always checking first for what the developer
explicitly typed, the developer has more control over ensuring
that another library isn't injected into the search order ahead
of it.

This commit swaps the order of the checks for prefix+name+suffix
and name, making the check for name first.