When HelperMethodFrame is created unwind context up to JIT code
On non-Windows OS current DAC implementation cannot upwind native stack
out-of-proc. That'w why when HelperMethodFrame is created at runtime
we pre-unwind stack up to JIT code, so DAC can just read it.
This has a potential performance impact on all calls from mscorlib to coreclr,
and we have to remove it as soon as DAC native stack unwind is implemented.