Try a fix for some frequently failing PAL tests
Some PAL tests are frequently failing in the CI. I noticed that the CI runs PAL tests specifying an output folder that is on a mounted share. The specified output folder is used as the working directory for all PAL tests. Perhaps it has something to do with that, but not sure yet.

Trying a few changes to way in which PAL are run, to see if the issue goes away:
- Use the default /tmp/PalTestOutput as the output directory
- Run each test in its own folder. Many PAL tests don't clean up after themselves, and create/use the same file/directory names in the current directory.
- Copy the output folder to the specified output folder at the end

Fix #1561