Fix string literals being deleted too early, use a proper Crst type
The function called by the LoaderAllocator finalizer was doing cleanup
too early and deleted string literals while other assemblies still had
access to them. This moves the cleanup into the LoaderAllocator GC so
it only gets deleted when it is no longer referenced.

The nested LoaderAllocator references were also being released too
early so this will skip it in the LoaderAllocator finalizer and do it
in the AssemblyLoadContext finalizer for collectible assemblies.