Easy switch between arm & arm64 for cross compile
If you define the ROOTFS_DIR environment variable, it is not easy
to switch compilation from arm to arm64 (and conversely) without
changing its value. This commit introduces a breaking change in the
meaning of the ROOTFS_DIR environment variable to not include the
architecture, it will be appended automatically by the build scripts.
It will make it also easier to share the root fs with CoreRT cross
compilation.