Add debugger launch to dbgshim for xplat.
Add the InvokeOnRuntimeStartup API to dbgshim. Executes the callback when the
coreclr runtime starts. The callback is passed the proper ICorDebug instance
for the version of the runtime. This API works for launch and attach (and even
the attach scenario if the runtime hasn't been loaded yet) on both xplat and
Windows.

If the coreclr runtime is already loaded in the process (the coreclr module
was found in the process), then ICorDebug instance is created and the callback
is invoked immediately on this API's calling thread.

If the runtime isn't found, a separate thread is created and the callback is
executed when the runtime startup is detected on this separate thread and this
API returns immediately.  The callback is invoke when the runtime is loaded and
initialized. The runtime is blocked during initialization until the callback
returns on Windows or when DebugActiveProcess is called on target process on
*unix/xplat.

Only the first coreclr module instance found in the target process is currently
supported.

HRESULT
InvokeOnRuntimeStartup(
    __in DWORD dwProcessId,
    __in PSTARTUP_CALLBACK pfnCallback,
    __in PVOID parameter)

Added PAL support functions PAL_InvokeOnRuntimeStartup and PAL_NotifyRuntimeStarted.

Enable WaitForSessionToOpen for the LS/debuggee to be used to
wait for the debugger connect on Launch.

Cleanup process code; remove now useless CProcSharedData.

Added "PROCESS" PAL trace type.