Fix an omission to treat as having side effects intrinsics that are modeled as calls.
Coverting some special calls to intrinsics for the purposes of the optimizers, so they can be better optimized. Such call
is object.GetType(). The new intrinsic is marked as no-side-effect if the
result of the call is not used the optimizer decided to eliminate the call
altogether. In caseswhere object is null, the NullReferenceException is
removed.

This fix adds the necessary code to treat such intrinsics as always having
side effects.