Modify Array.BinarySearch to return the first matching element
The BinarySearch methods used by the Array class perform a tradtional iterative search to find the desired value and quit iteration when that value is found. This commit modifies those searches to instead perform a deferred termination search that has the effect of returning the matching value in the array that is at the lowest index instead of returning whichever value is found first. The negative affect of this is that in situations where the value is discovered early the search must complete as if it wasn't found.