Ensure CreateAdjustmentRule is compiled into mscorlib.
Legacy SelfTest/SelfHost tests are currently calling CreateAdjustmentRule using reflection. When changes were made to no longer have an upstream caller of this method, the tests started failing because BclWriter was removing the method from mscorlib.

To fix the tests, and in case other external callers are invoking this method using reflection, add this member to model.xml to ensure BclWriter leaves it in mscorlib.