Fix for issue 2757.
This PR implements changes to the emitter and codegen to support gcinfo
generation for the second return register of 2 register returned structs.

It adds a testcase that tests the scenario.

It also generates the proper IR for GT_RETURN of implicit RetBuf methods.
Before the codegen was generating storing the address of the RetBuf in
RAX explicitly. With these changes the type of the method return is changed to BYREF
and the IR guarantees the right code is generated.

Some API cleanup:
Refactored the way multi-register GC return tyes are passed to the emitter
infrastructure to use a struct instead 2 parameters.
Refactored the genEmitCall signature to have a singe argument - a struct
that wraps all the existing arguments.