Fix for issue 2757.
This PR implements changes to the emitter and codegen to support gcinfo
generation for the second return register of 2 register returned structs.

It adds a testcase that tests the scenario.

It also generates the proper IR for GT_RETURN of implicit RetBuf methods.
Before the codegen was generating storing the address of the RetBuf in
RAX explicitly. With these changes the type of the method return is changed to BYREF
and the IR guarantees the right code is generated.