Remove the need for ResGen during build
Today we need a ResGen that supports preprocessor directives because we
use this in mscorlib.txt in order to cut out resources that only are
used in Desktop. The need for a ResGen made us pull in mono on
non-windows platforms.

I've added a task to buildtools to do a simple preprocessing of a
file. We use this to process the ifdef's in mscorlib.txt, generate a new
mscorlib.txt and then use the normal MSBuild system for using a text
file as resources.