Focusing on a proper integer rounding
As rightly pointed out via comments, the original code (i.e., outputing 123 when rounding 12345 to 3) doesn't represent a proper integer rounding and that's why it shouldn't be included in Math.Round (+ int.Round).

This new version does perform a proper integer rounding (i.e., 12300 in the aforementioned example) and can rightfully be included in both int.Round and in the Math.Round overloads taking int arguments. The previous functionality will be included in the new Truncate2 method; nevertheless, I have preferred to delete it until having a version accounting for all the types (also for the decimal ones).