Add getReadyToRunDelegateHelper to JitInterface
The existing approach to constructing delegates with the general purpose
getReadyToRunHelper has limitations for fully static construction of the
delegate. In particular, EE needs to know the delegate type along with
the method that the delegate is getting created for so that it can
distinguish between open/closed delegates at compile time.

This adds a new method to JIT-EE interface to allow that. The EE
implementation on the CoreCLR side doesn't use the new parameter right
now, but it might in the future.

It also has the nice property of not asking for a R2R helper unless
really needed.