Fix a bug in Compiler::fgMakeEmbeddedStmt when looking up for statement to be embedded.
This is a long standing bug in RyuJit.
The isue is that there are cases where we will not properly find a
statement that is to be made embedded in the list of existing statements,
starting with the parent statement.
In this particular case the code obtains the first node of the statement
that is to be made embedded (the first statement of the parameter tree), but doesn't do that for the first statement
of the parentStatement list.
The statement to be embedded is a call with controlExpr, so
getFirstNode(tree) includes the controlExpr. The
parentStmt->gtStmt.gtStmtList is the call itself, but since no
gtFirstNode() is called on it, the controlExpr is not included, preventing
to find properly the tree that is to be embedded in the parentStatement
list.

Fixes issue 3618.