Use current stack trace in ExceptionDispatchInfo when Exception lacks one
When implementing a component based on Tasks and wrapping a native asynchronous API, it's common to have code that manually completes tasks using TaskCompletionSources, e.g.
```C#
var tcs = new TaskCompletionSource<int>();
InvokeSomeNativeApi(r => tcs.SetResult(r), hr => tcs.SetException(GetExceptionFromHr(hr)));
return tcs.Task;
```
When an error occurs, typically an Exception is created for that native error, but the Exception isn't thrown and thus the Exception object doesn't contain a good stack trace.  Instead, the Exception is stored into the TaskCompletionSource with SetException.  SetException in turn uses ExceptionDispatchInfo.Capture to extract the relevant data from the Exception so that it can be restored and appended later when the exception is propagated.  Since there's no stack trace stored in the exception, when the Exception gets thrown later via ExceptionDispatchInfo.Throw, there's nothing for the EDI to append and the propagated exception lacks any information about the original source location.

This commit attempts to mitigate the problem by having the ExceptionDispatchInfo use the stack trace at the point of Capture if there's no other stack trace information already stored in the Exception.  Since it's typical for the Exception in such situations to be constructed in the same general vicinity as where it's then Capture'd, this provides more relevant information than an empty stack trace.