Specify the signedness of 'char' in HFA testcases
In C#, sbyte always represents 8-bit signed interger.

However, the signedness of char in C is undefined, and the default
signedness of char in ARM is unsigned.

The native code for HFA test does not specify the signedness of char,
and thus the native code assumes that it is unsigned, but the managed
code assumes that it is signed, which leads to unittest failure in ARM.

This commit specifies the signedness of 'char' in order to fix #4639.