Make sure explicit tail call result is never stored on stack.
For purposes of debugging the last call of a method is normally spilled on
the stack to ensure the value ius examinable by the debugger when compiling
for debug configuration.

For explicit tail calls this breaks an assumption in the tail call optimizer
that the IR node after the explicit tail call node is a return.

Make sure if the call is marked as tail call, no spilling or any other
nodes are generated between the call and the return - the explicit tail
call will be conveted to a jmp, so no need to store the result in a temp
var on the stack.

Also modified the ilprojfile for this test to make sure it is always
compiled with optimizations on - an assumption in the tests.

Also, made a minor fix in the test itself.

Fixes #2329.