Handle tail calls with following GT_NOPS ane GT_RETURN.
For purposes of debugging the last call of a method is normally spilled on
the stack to ensure the value is examinable by the debugger when compiling
for debug configuration. The GT_ASG node (that does the spilling) is
followed
by a GT_COMMA. The GT_COMMA node in this case is not needed, so this
change is
making sure it is not being generated. The codegen generates an GT_NO_OP
node after the
call. A tail callable patter could contain any number of NOPs between the
tail call and the return.
The tail call of the optimizer is not handling this case and this change
also adresses that issue.

Also modified the ilproj file for this test to make sure it is always
compiled with optimizations on - an assumption in the tests.

Also, made a minor fix in the test itself.

Fixes #2329.