Use "debian.8-x64" as the debian RID.
The RID we had been using was not correct (it's something we made up
that the CLI would never generate. We should move away from it some we
can remove it from the list of rids in our rid graph).

I generated this commit via a tool that updated every project.json with
a debian.8.2 rid in it.  You'll notice some sorting and formating
changes because of that.