Fix Guid's GetHashCode to work better (not ignore parts of it)
A common pattern is that the last bits of the GUID are the most unique, however our current hash code ignores most of them.
Fix this to simply XOR all bits in the GUID 32 bits at a time.   This is more efficient and a better hash than what we have currently.