CoreDisTools Package: Make minor version generic
Change the version requirement of CoreDisTools package
from 1.0.1-prerelease-00001 to 1.0.1-prerelease-*
so that the test infrastructure picks up the latest
available version.

If only the 1.0.1-prerelease-xxxxx number changes,
there is no change in the library API. There may be
a change in the implementation.

The current need for this change is that we need to pick up
a new version of the CoreDisTools package (with an
implementation fix) for X86 testing.