Make Reflection's GetMethods() return methods in declaration order consistantly.
For classes with less than 64 methods, the GetMethods() call returns the methods in the order
of delcaration.   However internally in the runtime we read methods in chunks and put
them on a linked list, which means that for classes with more than this number of methods
the order gets scrambled.

There are classes like System.Diagnostic.Tracing.EventSource that depend on reflection
returning the methods in order of declaration IN CERTAIN SCENARIOS, and this does work
today for classes with less than 64 methods.     However we are now seeing cases where people
are relying on this scenario with classes of more than 64 methods.

It is pretty trivial to fix this so that we preserve the order of the methods.   Basicaly we
reverse the list of chunks after we have generated it, which restores the order.

This fix can be viewed as a compatibility fix, but in general preserving delclaration order
is a useful, because it is a natural way of resolving conflicts (e.g. last one wins), and
there is no other way to get this information from reflection unless we expose it in
GetMethods().     we are 95% of the way there (it works in chunks), this simple
change pushes us to 100% (for statically declared methods).