Fix backwards compatibility bug: RC2 dbgshim debugging RTM runtime.
Issue #5569

Debugging the RTM runtime with the current VSCode C# extension (RC2 dbgshim) is broken. The target app just runs to completion and any initial breakpoints are not hit.

If we do NOT take this then:
• On the day of RTM, anyone installing VS Code or VS Docker tooling will instantly find that it has stopped working for any existing RC projects until they update to RTM.
• It will not be possible to launch using VS Code until we are ready to start producing RTM builds (when tooling goes into escrow). Folks who want this in that time window can manually get things working by copying over their dbgshim.

The fix is to change the name of the "continue" semaphore used in the launch/attach handshake by the RTM runtime to match the one the RC2 dbgshim creates.