Updates tryrun.cmake for ARM / ARM (softfp)
While fixing #5316 (PR #5317, PR #5323), all the features related with PROCFS
are disabled in ARM and ARM (softfp) cross build.

As a result, the collision avodance logic in GetProcessIdDisambiguationKey is disabled,
which leads to random coreclr_initialize failures
(CLR failed to initialize itself when the previous execution of CLR with the same PID
abnormally terminates)

This commit tries to re-enable the relevant features for ARM and ARM (softfp) cross-build.

Since the ARM/Linux emulator (which is the current reference in use) supports
/proc/self/stat, /proc/self/status, /proc/self/maps, this commit re-enables
three relevant features (HAVE_PROCFS_STAT, HAVE_PROCFS_STATUS, HAVE_PROCFS_MAPS).

This commit tries to fix #5583.